Visual Studio 2005 のコンパイル設定を追加。 また、Visual Studio 2005 で、コンパイルが通るようにした。

+++ mist/filter/decomposition.h
@@ -463,19 +463,19 @@ typename Array1::size_type figure_decomposition( const Array1 &in, Array2 &out, 
 					{
 						size_type pz = k + z;
 						if( pz < 0 || pz >= in.depth( ) ) continue;
-						double kk = k * k;
+						double kk = static_cast< double >( k * k );
 
 						for( difference_type j = -ry ; j <= ry ; j++ )
 						{
 							size_type py = j + y;
 							if( py < 0 || py >= in.height( ) ) continue;
-							double jj = j * j;
+							double jj = static_cast< double >( j * j );
 
 							for( difference_type i = -rx ; i <= rx ; i++ )
 							{
 								size_type px = i + x;
 								if( px < 0 || px >= in.width( ) ) continue;
-								double rR = i * i + jj + kk;
+								double rR = static_cast< double >( i * i + jj + kk );
 
 								size_type cl = static_cast< size_type >( out( px, py, pz ) );
 								if( cl != 0 && min > rR )
